extract user types
