+++ packages/client/src/api/contexts/RatingContext.tsx
@@ -2,7 +2,7 @@ import { useToast } from '@chakra-ui/react'
 import { useQuery } from '@tanstack/react-query'
 import { createContext, PropsWithChildren, useEffect, useState } from 'react'
 import { useLocation, useNavigate, useSearchParams } from 'react-router-dom'
-import { functionAxios } from '../../util/initAxios'
+import { functionAxios } from '../../util/axiosConfig'
 import { PATHS } from '../../util/paths'
 import { useSubmitRatingMutation } from '../hooks/ratingHooks'
 import { CategoryWithCriteria, DbStage, EventRatingInfo } from '@pontozo/types'
