+++ packages/client/src/api/hooks/ratingHooks.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQuery } from '@tanstack/react-query'
 import { FUNC_HOST } from '../../util/environment'
-import { functionAxios } from '../../util/initAxios'
+import { functionAxios } from '../../util/axiosConfig'
 import { CriterionRating, EventRatingInfo, GetCriterionRatings, GetEventRating, CreateEventRating } from '@pontozo/types'
 
 export const useStartRatingMutation = () => {
