--- packages/client/src/util/transformUser.ts
@@ -1,7 +1,0 @@
-import { User, UserPreview } from '../api/model/user'
-
-export const transformUser = (u: User): UserPreview => ({
-  userDOB: u.szul_dat,
-  userFullName: u.vezeteknev + ' ' + u.keresztnev,
-  userId: u.szemely_id,
-})
