--- packages/functions/src/functions/auth/types/MtfszUser.ts
@@ -1,13 +1,0 @@
-export interface MtfszUser {
-  felhasznalo_id: number
-  szemely_id: number
-  keresztnev: string
-  vezeteknev: string
-  nev: string
-  email: string
-  nem: string
-  titulus: string
-  szul_dat: string
-  username: string
-  v_nytsz: string
-}
