--- packages/functions/src/functions/auth/types/PontozoUser.ts
@@ -1,6 +1,0 @@
-import { UserRole } from '../../../typeorm/entities/UserRoleAssignment'
-import { MtfszUser } from './MtfszUser'
-
-export interface PontozoUser extends MtfszUser {
-  roles: UserRole[]
-}
