--- packages/functions/src/functions/auth/types/Token.ts
@@ -1,6 +1,0 @@
-export interface Token {
-  token_type: string
-  expires_in: number
-  access_token: string
-  refresh_token: string
-}
