+++ packages/types/src/index.ts
@@ -7,6 +7,7 @@ export * from './lib/criterionRatings'
 export * from './lib/eventRatings'
 export * from './lib/seasons'
 export * from './lib/uras'
+export * from './lib/users'
 
 export type CreateResponse = {
   id: number
@@ -14,4 +15,9 @@ export type CreateResponse = {
 
 export type EntityWithEditableIndicator<T> = T & {
   editable: boolean
+}
+
+export type MtfszFetchResult<T> = {
+  result: T[]
+  page_size: number
 }
