+++ packages/types/src/lib/users.ts
@@ -1,6 +1,6 @@
-import { UserRole } from '@pontozo/types'
+import { UserRole } from './uras'
 
-export interface User {
+export interface MtfszUser {
   felhasznalo_id: number
   szemely_id: number
   keresztnev: string
@@ -14,7 +14,7 @@ export interface User {
   v_nytsz: string
 }
 
-export interface UserDetails extends User {
+export interface DbUser extends MtfszUser {
   roles: UserRole[]
 }
 
@@ -22,4 +22,11 @@ export interface UserPreview {
   userId: number
   userFullName: string
   userDOB: string
+}
+
+export interface MtfszToken {
+  token_type: string
+  expires_in: number
+  access_token: string
+  refresh_token: string
 }
