+++ packages/types/tsconfig.lib.json
@@ -1,10 +1,10 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "outDir": "../../dist",
+    "outDir": "./dist",
     "declaration": true,
     "types": ["node"]
   },
-  "include": ["src/**/*.ts"],
-  "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"]
+  "include": ["./src/**/*.ts"],
+  "exclude": ["./jest.config.ts", "./src/**/*.spec.ts", "./src/**/*.test.ts"]
 }
